Minor refactor of pmg scripts.
