+++ pymatgen/cli/pmg.py
@@ -21,7 +21,7 @@ from pymatgen.io.vasp import Incar
 
 from pymatgen.cli.pmg_analyze import parse_vasp
 from pymatgen.cli.pmg_config import configure_pmg
-from pymatgen.cli.pmg_generate_potcar import generate_potcar
+from pymatgen.cli.pmg_potcar import generate_potcar
 from pymatgen.cli.pmg_plot import plot
 from pymatgen.cli.pmg_structure import analyze_structures
 from pymatgen.cli.pmg_query import do_query
